Rename and move over convertIsoToDateObj
