+++ cypress_shared/pages/page.ts
@@ -1,5 +1,5 @@
 import errorLookups from '../../server/i18n/en/errors.json'
-import { convertDateString } from '../../server/utils/utils'
+import { DateFormats } from '../../server/utils/dateUtils'
 
 export type PageElement = Cypress.Chainable<JQuery>
 
@@ -63,8 +63,7 @@ export default abstract class Page {
   }
 
   completeDateInputs(prefix: string, date: string): void {
-    const parsedDate = convertDateString(date)
-
+    const parsedDate = DateFormats.convertIsoToDateObj(date)
     cy.get(`#${prefix}-day`).type(parsedDate.getDate().toString())
     cy.get(`#${prefix}-month`).type(`${parsedDate.getMonth() + 1}`)
     cy.get(`#${prefix}-year`).type(parsedDate.getFullYear().toString())
