+++ server/utils/dateUtils.test.ts
@@ -1,6 +1,26 @@
 import { DateFormats, InvalidDateStringError, dateAndTimeInputsAreValidDates } from './dateUtils'
 
 describe('DateFormats', () => {
+  describe('convertIsoToDateObj', () => {
+    it('converts a ISO8601 date string', () => {
+      const date = '2022-11-11T00:00:00.000Z'
+
+      expect(DateFormats.convertIsoToDateObj(date)).toEqual(new Date(2022, 10, 11))
+    })
+
+    it('raises an error if the date is not a valid ISO8601 date string', () => {
+      const date = '23/11/2022'
+
+      expect(() => DateFormats.convertIsoToDateObj(date)).toThrow(new InvalidDateStringError(`Invalid Date: ${date}`))
+    })
+
+    it('raises an error if the date is not a date string', () => {
+      const date = 'NOT A DATE'
+
+      expect(() => DateFormats.convertIsoToDateObj(date)).toThrow(new InvalidDateStringError(`Invalid Date: ${date}`))
+    })
+  })
+
   describe('isoDateToUIDate', () => {
     it('converts a ISO8601 date string to a GOV.UK formatted date', () => {
       const date = '2022-11-11T00:00:00.000Z'
