+++ server/utils/dateUtils.ts
@@ -29,6 +29,22 @@ export class DateFormats {
   }
 
   /**
+   * Converts an ISO8601 datetime string into a Javascript Date object.
+   * @param date An ISO8601 datetime string
+   * @returns A Date object
+   * @throws {InvalidDateStringError} If the string is not a valid ISO8601 datetime string
+   */
+  static convertIsoToDateObj(date: string) {
+    const parsedDate = parseISO(date)
+
+    if (Number.isNaN(parsedDate.getTime())) {
+      throw new InvalidDateStringError(`Invalid Date: ${date}`)
+    }
+
+    return parsedDate
+  }
+
+  /**
    * @param isoDate an ISO date string.
    * @returns the date in the to be shown in the UI: "Thursday, 20 December 2012".
    */
@@ -36,3 +52,5 @@ export class DateFormats {
     return DateFormats.dateObjtoUIDate(DateFormats.convertIsoToDateObj(isoDate))
   }
 }
+
+export class InvalidDateStringError extends Error {}
